requirements.txt
# Core functionality
python-dotenv
chromadb
langgraph
psycopg2-binary
pytz
boto3

# LlamaIndex Core and specific integrations
llama-index-core
llama-index-vector-stores-chroma
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-readers-file

# Google Client Libraries for Gmail API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib